+++ Transaction.java
@@ -1,10 +1,10 @@
 public class Transaction {
 	private String username;
 	private TransactionType type;
-	private int amount;
+	private double amount;
 	private boolean success;
 
-	public Transaction(String username, TransactionType type, int amount, boolean success) {
+	public Transaction(String username, TransactionType type, double amount, boolean success) {
 		this.username = username;
 		this.type = type;
 		this.amount = amount;
